# perf: reduces bazel query output