@iamricard 2023-12-11
> Thanks, LGTM! #312 should fix the CI failures. Once that merges, can you rebase on top?

done!