@iamricard 2023-12-11
thanks for merging @coeuvre 🙇🏼 do you have an idea of when y'all are likely to cut a new release? i'd like to let folks know at datadog they can go back to using the published extension 🎉