@daivinhtran 2023-12-20
@iamricard I'm afraid this optimization is breaking CodeLens (or maybe only on the bazel version my repo is using).

I tried reverting the proto flags and CodeLens worked as normal. See the [screencast](https://github.com/bazelbuild/vscode-bazel/assets/13268391/e58b4839-34d0-4246-9d6c-ffe887ad1cd9) when the flags are used:

- Clicking on the targets in Bazel Build Targets no longer works.
- CodeLens feature that allows navigating to a target no longer works.
- The tooltips for build command now all shows at the top of the BUILD file instead of being inline with the targets.